Move `/data` out of `/assets`
