+++ components/ReadTheLetter.vue
@@ -3,7 +3,7 @@
     <h2 class="text-center">
       Read the Letter
     </h2>
-    <div class="sml-push-y2 med-push-y3" v-html="letterText" />
+    <div class="sml-push-y2 med-push-y3" v-html="letterText"></div>
   </div>
 </template>
